Use full path to 'find'
This is just an ad-hoc fix to make aab work on Windows with a Unix
version of find. I have Chocolatey's finutils package in PATH but Windows'
unrelated `find` command was getting executed despite that. This fixes
it by getting the full command path using shutil.which.
Note that this fails if `find` is in a folder with spaces in it.